Extract models.internal.dart
