+++ lib/app_store_screenshots.dart
@@ -3,4 +3,4 @@ library app_store_screenshots;
 export 'src/generate_app_icons.dart';
 export 'src/generate_google_play_assets.dart';
 export 'src/generate_screenshots.dart';
-export 'src/models.dart' hide DeviceTypeExtensions;
+export 'src/models.dart';
